Core: Updated to 11.2.7
Port From (https://github.com/TrinityCore/TrinityCore/commit/bd2b12fa36830c6b12e124359f46ac974d554080)

+++ Source/Game/Networking/Packets/TalentPackets.cs
@@ -31,7 +31,7 @@ namespace Game.Networking.Packets
                     talent.Write(_worldPacket);
 
                 foreach (uint talent in talentGroupInfo.GlyphIDs)
-                    _worldPacket.WriteUInt16((ushort)talent);
+                    _worldPacket.WriteUInt32(talent);
             }
         }
